[refactor] post, edit 페이지 리팩토링
