--- src/app/(community)/community-list/page.tsx
@@ -1,17 +1,0 @@
-// CommunityPage.tsx
-'use client';
-
-import CommunityMain from '../(components)/CommunityMain';
-import { Suspense } from 'react';
-
-const CommunityPage = () => {
-  return (
-    <div>
-      <Suspense>
-        <CommunityMain />
-      </Suspense>
-    </div>
-  );
-};
-
-export default CommunityPage;
