--- src/app/(community)/community-post/page.tsx
@@ -1,11 +1,0 @@
-import PostForm from '../(components)/PostForm';
-
-const PostPage = () => {
-  return (
-    <>
-      <PostForm />
-    </>
-  );
-};
-
-export default PostPage;
