+++ src/app/community-list/[category]/page.tsx
@@ -1,0 +1,16 @@
+import React, { Suspense } from 'react';
+import CommunityMain from './CommunityMain';
+
+type Props = {};
+
+const Page = (props: Props) => {
+  return (
+    <div>
+      <Suspense>
+        <CommunityMain />
+      </Suspense>
+    </div>
+  );
+};
+
+export default Page;
