+++ src/app/community-list/page.tsx
@@ -1,0 +1,7 @@
+import { redirect } from 'next/navigation';
+
+const CommunityPage = () => {
+  redirect('/community-list/전체');
+};
+
+export default CommunityPage;
