+++ src/app/community-list/write/layout.tsx
@@ -1,0 +1,21 @@
+import CategorySelector from '@/app/(components)/CategorySelector';
+import React from 'react';
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const Layout = ({ children }: Props) => {
+  return (
+    <main className="grid grid-cols-[16%_84%] h-[84vh]">
+      <section className="flex flex-col justify-between h-[84vh] bg-bgColor1 border-r-2 border-solid border-pointColor1">
+        <div>
+          <CategorySelector categoryNow={''} />
+        </div>
+      </section>
+      {children}
+    </main>
+  );
+};
+
+export default Layout;
